Authenticate user from Onadata db when making requests to import the form url into KPI and also when fetching form information from Onadata.

+++ kpi/signals.py
@@ -43,17 +43,18 @@ def save_kobocat_user(sender, instance, created, raw, **kwargs):
     grant all KoBoCAT model-level permissions for the content types listed in
     `settings.KOBOCAT_DEFAULT_PERMISSION_CONTENT_TYPES`
     """
-    if not settings.TESTING:
-        KobocatUser.sync(instance)
+    if False:
+        if not settings.TESTING:
+            KobocatUser.sync(instance)
 
-        if created:
-            # FIXME: If this fails, the next attempt results in
-            #   IntegrityError: duplicate key value violates unique constraint
-            #   "auth_user_username_key"
-            # and decorating this function with `transaction.atomic` doesn't
-            # seem to help. We should roll back the KC user creation if
-            # assigning model-level permissions fails
-            grant_kc_model_level_perms(instance)
+            if created:
+                # FIXME: If this fails, the next attempt results in
+                #   IntegrityError: duplicate key value violates unique constraint
+                #   "auth_user_username_key"
+                # and decorating this function with `transaction.atomic` doesn't
+                # seem to help. We should roll back the KC user creation if
+                # assigning model-level permissions fails
+                grant_kc_model_level_perms(instance)
 
 
 @receiver(post_save, sender=Token)
@@ -61,8 +62,9 @@ def save_kobocat_token(sender, instance, **kwargs):
     """
     Sync AuthToken table between KPI and KC
     """
-    if not settings.TESTING:
-        KobocatToken.sync(instance)
+    if False:
+        if not settings.TESTING:
+            KobocatToken.sync(instance)
 
 
 @receiver(post_delete, sender=Token)
@@ -70,11 +72,12 @@ def delete_kobocat_token(sender, instance, **kwargs):
     """
     Delete corresponding record from KC AuthToken table
     """
-    if not settings.TESTING:
-        try:
-            KobocatToken.objects.get(pk=instance.pk).delete()
-        except KobocatToken.DoesNotExist:
-            pass
+    if False:
+        if not settings.TESTING:
+            try:
+                KobocatToken.objects.get(pk=instance.pk).delete()
+            except KobocatToken.DoesNotExist:
+                pass
 
 
 @receiver(post_save, sender=Tag)
